Dynamic Canvas Size
Added some functions to make it so that the size of the window is dependent on the size of the screen it's being viewed on! It's also capable of changing with the window's being resized, so it will never be too big or ridiculously small. I might want to add a limit to the dimension for the sake of lag prevention, though. I'll do that in the next commit!